backend/pyproject.toml
[tool.poetry]
name = "haeksim"
version = "0.1.0"
description = "LLM based problem solving evaluation system"
authors = ["chr-kim <chriankim@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
google-genai = "^1.36.0"
python-dotenv = "^1.0.1"
PyMuPDF = "^1.24.9"
streamlit = "^1.36.0"
requests = "^2.32.3"
pytest = "^7.0.0"
black = "^23.0.0"
sqlalchemy = {extras = ["aiosqlite"], version = "^2.0.43"}
databases = {extras = ["sqlite"], version = "^0.9.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"